core/translate: bind strict check types in bind module
Move strict CHECK column, function, and custom-type resolution into bind.rs with the rest of CHECK binding. Schema validation now delegates the complete recursive comparison analysis to the binder.

Validation: cargo check -p turso_core; cargo fmt --all -- --check
Tests: not run as requested
